fixes NPE
git-svn-id: https://server.mycore.de/svn/mycore/trunk@17735 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-iview2/src/main/java/org/mycore/frontend/iview2/MCRIView2Commands.java
@@ -230,7 +230,7 @@ public class MCRIView2Commands extends MCRAbstractCommands {
         File parent = file.getParentFile();
         if (file.exists())
             file.delete();
-        if (parent != null && parent.list().length == 0)
+        if (parent != null && parent.isDirectory() && parent.list().length == 0)
             deleteFileAndEmptyDirectories(parent);
     }
